refactor(asr): extract distribution fitting and covariance helpers
Move fit_eeg_distribution (+ histogram / robust-stats helpers) to
_distribution.py and the block/rolling covariance builders, robust aggregation,
and memory-budgeting helpers to _covariance.py. core.py re-exports them via
__all__; no behavior change.
